+++ tests/lib/js/static_interop_test/import/import_test.dart
@@ -5,17 +5,18 @@
 import 'dart:js_interop';
 
 import 'package:async_helper/async_helper.dart';
-import 'package:expect/minitest.dart'; // ignore: deprecated_member_use_from_same_package
+import 'package:expect/expect.dart';
 
-extension type Module(JSObject o) implements JSObject {
+extension type Module(JSObject _) implements JSObject {
   external String testModuleFunction();
 }
 
 void main() {
   asyncTest(() async {
     final module = Module(await importModule(
-            '/root_dart/tests/lib/js/static_interop_test/import/module.mjs')
+            '/root_dart/tests/lib/js/static_interop_test/import/module.mjs'
+                .toJS)
         .toDart);
-    expect(module.testModuleFunction(), 'success');
+    Expect.equals(module.testModuleFunction(), 'success');
   });
 }